requirement.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==1.4.6
pillow==10.3.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
